Added "Airbnb Clone" project
This commit adds the "Airbnb Clone" project that I completed as part of the React Basics course.
During the course, I gained valuable knowledge on how to work with props in React and how to
effectively map data from props to components. I applied these skills to develop a simple yet
dynamic website that retrieves data from an external file and populates the page with the data dynamically.